Fixed zookeeper quorum bug: BIGTOP-2467
The existing Bigtop puppet scripts write out the Zookeeper peers in such
a way that it is difficult to maintain a consistent id for each peer
across zoo.cfg files on different machines. This breaks Zookeeper,
because it expects peers to have a consistent id across machines.

This patch tweaks the puppet script so that it takes an explicit id in
addition to an ip address for each element that it writes out to the
config file, allowing the charm to majage the ids itself.