Puppet patch, and fixes for Zookeeper on Xenial. (#33)
Puppet patch contains pass-through for clientPortAddress, so that we can
bind Zookeeper to listen on a specfic interface.

As for the fixes ...

On Ubuntu Xenial, a Zookeeper package exists in universe, and we wind up
installing it rather than the one in Bigtop. This breaks many things.

In this change, we attempt to pin any given package to the Bigtop repo,
which fixes this, and should fix similar problems in the future.