replace smoke test config files with env
We currently allow charms to specify a smoke_test_configs layer opt,
which is supposed to be files that we source prior to running the
smoke test. This is hard to handle because we don't know if that
file will be an ini, a dotfile, or some other thing that we have to
grok and source.

Seems smarter for the smoke test itself to construct a dict of all
the environment stuff it needs and pass that into the smoke test
runner.

This does that. I do not believe anyone is using the config file
madness, and it certainly didn't do anything in the smoke test
runner function, so I feel like now's the time to get on the right
path.