ugh. more unit test tomfoolery.
Everybody calls Bigtop(). We cant just set bigtop_apt=None in __init__.
Duh.

Refactor this madness to set the actual apt value in __init__, yet
leave the logic in a (more aptly named) get_repo_url method so we can
still unit test this sucker.

This means we'll need to mock a version for the setUp unit test, which
is what I was trying to avoid in the first place... for reasons that
now escape me.