fail faster on our template reinstall action
Reinstall needs to be modified for each charm to do specific app handling
(minimally, to render a new site.yaml).

We want charm authors to leverage our template, but we don't want end
users to think that the default reinstall action is fully functional. Fail
fast if we detect that a user is attempting to run an unmodified reinstall
action.