bigtop repo url fixes
Revert pieces of #66. For bigtop 1.1.0 and 1.2.0, the correct repos
are still using bigtop-repos.s3. See the relevant bigtop.list at our
source of truth:

https://www.apache.org/dist/bigtop/

Set default release/series to 16.04/xenial. These are installable on
non-LTS releases, but are required to match the bigtop repo layout.

Remove useless checks for dist_name == 'ubuntu' since we fail fast
if someone tries to deploy a charm on non-ubuntu.

Kafka and non-intel are no longer available in the official repos
after 1.2. Ensure those cases are handled by setting the repo to the
correct CI repository.

Update the unit tests to deal with the above.